dominant color processing on a full image + Fixes (#1856)
* Fix - Retries are not done when connection is destructed

* Remove costly unpack and streamline code

* Update Flatbuffer client code

* Support for dominant color processing on a full image which is applied to all LEDs (#1853)

* Allow uni dominant enums for API

* Fix C&P issue

* Missing remote buttons' text

* Allow to provide additional details in error dialogue

* Ensure UI getConfig always get instances

* Stop API service when Hyperion is quitting

* Update UI that Single and Multiple instance commands are correctly supported

* Fix that effects are not created/deleted per instance

* Fix dangling reference

* Correct instance dependencies on API commaands

* Add failing command to Error dialog

* Fix - Return correct mapping type for a running instance

* Update changelog

* Cleanup Flatbuffer processing

* Use switched instance for API calls without instance provided

* Treat an empty instance array as "all instances" for multi-instance commands

* Fixed - Last update of an effect event is not removed in sources overview

* Translate no sources text element

* Fix CodeQL finding

* JSON API clean-up

* Http-Server: Support Cross-Origin Resource Sharing (CORS) (#1496)

* Dress CodeQL Findings (2)

+++ include/utils/ImageResampler.h
@@ -17,7 +17,7 @@ public:
 	void setCropping(int cropLeft, int cropRight, int cropTop, int cropBottom);
 	void setVideoMode(VideoMode mode) { _videoMode = mode; }
 	void setFlipMode(FlipMode mode) { _flipMode = mode; }
-	void processImage(const uint8_t * data, int width, int height, int lineLength, PixelFormat pixelFormat, Image<ColorRgb> & outputImage) const;
+	void processImage(const uint8_t * data, int width, int height, size_t lineLength, PixelFormat pixelFormat, Image<ColorRgb> & outputImage) const;
 
 private:
 	int _horizontalDecimation;
